+++ examples/demo-junit/build.gradle
@@ -2,6 +2,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
     testImplementation project(':proxito-core')
+    testImplementation project(':proxito-junit')
     testImplementation project(':proxito-wiremock')
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
