+++ proxito-core/src/main/kotlin/io/fries/proxito/core/context/ProxitoContext.kt
@@ -3,10 +3,3 @@ package io.fries.proxito.core.context
 fun interface ProxitoContext {
     fun path(): String
 }
-
-data class JunitContext(
-    private val className: String,
-    private val testName: String,
-) : ProxitoContext {
-    override fun path(): String = "$className/$testName"
-}
