+++ proxito-junit/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation project(':proxito-core')
+    implementation 'org.springframework.boot:spring-boot-starter-test'
+}
