+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'proxito'
 
 include ':proxito-core'
+include 'proxito-junit'
 include ':proxito-wiremock'
 
 include ':examples:demo-junit'
